+++ libs/f-flow/src/f-connection/f-snap-connection/index.ts
@@ -1,1 +1,2 @@
 export * from './f-snap-connection.component';
+export * from './f-snap-target-change-event';